I61 no join requests stored (#62)
* Removed redundant join requests.

* Updated example with dummy oracles.

* All examples tested.

* Clippy fixes.

* Fmt

* Oracle answers are bundled with the result on-chain.

Thus the oracle operators only have to submit a true/false result
instead of compiling a whole struct.

+++ guild-network-test-runtime/src/lib.rs
@@ -25,7 +25,7 @@ macro_rules! test_runtime {
 
         impl pallet_randomness_collective_flip::Config for TestRuntime {}
 
-        pub type AccountId = u128;
+        pub type AccountId = u64;
         pub type Balance = u64;
         pub type Block = frame_system::mocking::MockBlock<TestRuntime>;
         pub type BlockNumber = u64;
